Update rMutation.sage
Added a way to just pass a linear order instead of the whole GIM.
Tested on every example in section 2, and it works! The only one it had trouble with was example 2.7, but I believe that calculation is off slightly in the paper. There should be no r_2's in r_4^w.